Fix compile warnings
- Fix warnings related to wrong types
- Fix compile warnings from unused vars
- Fix compile warnings array overflow
- Fix compile warning comment marker
- Remove compile warning re ftime
- Fix Wreorder warnings

+++ src/fsxwmech.cpp
@@ -839,7 +839,7 @@ double Mechanix::spreadrate(long slope, double windspd, int fuel)
      if(ndead>0)
           ndead=4;
 
-     double nclas[2]={ndead,nlive};  // # of dead & live fuel classes
+     double nclas[2]={(double) ndead, (double) nlive};  // # of dead & live fuel classes
     	double xmext[2]={newfuel.xmext, 0};
 
     	double load[4][2]=			     // tons per acre, later converted to lb/ft2
@@ -867,10 +867,10 @@ double Mechanix::spreadrate(long slope, double windspd, int fuel)
      //-------------------------------------------------------------------------
 
 	double sav[4][2]=			     // 1/ft
-	{	{newfuel.sav1, newfuel.savlh},
-		{109.0, newfuel.savlw},
+             {	{(double) newfuel.sav1, (double) newfuel.savlh},
+		{109.0, (double) newfuel.savlw},
 		{30.0, 0.0},
-		{newfuel.savlh, 0.0},
+		{(double) newfuel.savlh, 0.0},
 	};
 
 	double heat[4][2]=
